Replace magic numbers
The previous implementation was awful. Trajectory keys were used for
arthimetic (!) in calculating record sizes. Furthermore, an undocumented
key (3) appears for some kind of compressed trajectory output (using
netcdf?). The relevant areas of code are reached using goto (!)
statements, it is unclear whether this is supported or works.

The magic number have been replaced with parameters keys and a tbp of
the trajectory type has been created to return the corresponding keys
used in the trajectory files and control file.