Printing the mutable parameters of the KIM portable models
Addition of the KIM model type in the 'kim_setup', where KIM
model has been added for the addition of the KIM simulator
and its future support in DL_POLY. First, it checks if the
model is a portable or a simulator model.

Calling 'kim_setup' would report the mutable parameters
of the KIM portable models (PM) including parameter's name,
data type, and extent.