Update the kim module according to kim_init & kim_interactions
- Adding the kim_interactions function that defines a list of N
species, and defines a mapping between atom types in DL_POLY
to the available species in the OpenKIM IM
- Updating the kim_energy_and_forces accordingly to work with the new
update. This optimization improve performance to 25% and up.
Now we are using the array of unique_species_code to assign the species
code from kim api and avoid extra function calls.