Updating a KIM model specification
- The `kim_init` keyword followed bys the model name of the interactomic
model to be used in the simulation. It sets the IM model that will be
used and may issue additional update to the DL_POLY settings that are
required for running the simulation.
- The `kim_interactions` keyword must be preceded by a `kim_init` keyword
and a `name-of-molecule`, `nummols`, and `atoms`, etc  keywords that define
the system. The `kim_interactions` keyword must be followed by a list of N
unique species, which defines a mapping between atom types in DL_POLY to
the available species in the OpenKIM IM